// client/src/state/action_types/gameActionTypes.ts
enum gameActionType {
  NEW_GAME = "NEW GAME",
  MOVE = "MOVE",
  UPDATE_GAME = "UPDATE GAME",
  ROTATE_BOARD = "ROTATE BOARD",
  GAME_OVER = "GAME OVER",
  OWN_DRAW_OFFER_ON = "OWN_DRAW_OFFER_ON",
  OWN_DRAW_OFFER_OFF = "OWN_DRAW_OFFER_OFF",
  OPONENT_DRAW_OFFER_ON = "OPONENT_DRAW_OFFER_ON",
  OPONENT_DRAW_OFFER_OFF = "OPONENT_DRAW_OFFER_OFF",
  RESET = "RESET",
}

export default gameActionType;